pyproject.toml
[project]
name = "mcp-shioaji"
version = "0.1.0"
description = "MCP server for Shioaji"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.5.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "shioaji[speed]>=1.2.5",
]
